Moved helper method as per @littleforest's suggestion.
