+++ spec/rails_helper.rb
@@ -27,6 +27,7 @@ RSpec.configure do |config|
   config.include SessionHelper, type: :view
   config.include SessionHelper, type: :request
   config.include Organizational, type: :view
+  config.include CssHelpers
   config.after do
     Warden.test_reset!
   end
@@ -47,17 +48,4 @@ RSpec.configure do |config|
   config.example_status_persistence_file_path = "#{::Rails.root}/tmp/persistent_examples.txt"
 
   config.filter_rails_from_backtrace!
-end
-
-
-# Returns a hash of _specified_ (as opposed to runtime-adjusted/overridden) CSS attributes for an element
-def specified_style_attributes(capybara_element)
-  style_string = capybara_element['style']
-  attribute_strings = style_string.split(';')
-  attribute_strings.each_with_object({}) do |string, style_hash|
-    first_colon_position = string.index(':')
-    key = string[0...first_colon_position].strip
-    value = string[(first_colon_position+1)..-1].strip
-    style_hash[key] = value
-  end
 end
